Restructure the ui
