+++ ui/index.html
@@ -1,0 +1,83 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>ZtoApi - OpenAI-compatible API proxy</title>
+    <link rel="stylesheet" href="/ui/index.css">
+</head>
+<body>
+    <div class="container">
+        <header>
+            <h1>ZtoApi</h1>
+            <div class="subtitle">OpenAI-compatible proxy for Z.ai GLM-4.5</div>
+            <p>A high-performance, easy-to-deploy API proxy that allows you to use OpenAI-compatible formats to access Z.ai's GLM-4.5 models.</p>
+        </header>
+
+        <div class="links">
+            <div class="link-card">
+                <h3>📖 API Docs</h3>
+                <p>View full API documentation and usage examples.</p>
+                <a href="/docs">View Docs</a>
+            </div>
+
+            <div class="link-card">
+                <h3>📊 Dashboard</h3>
+                <p>Real-time monitoring of API calls, statistics, and performance.</p>
+                <a href="/dashboard">View Dashboard</a>
+            </div>
+
+            <div class="link-card">
+                <h3>🤖 Models</h3>
+                <p>See the available AI models and their details.</p>
+                <a href="/v1/models">View Models</a>
+            </div>
+        </div>
+
+        <div class="features">
+            <h2>Features</h2>
+            <div class="feature-list">
+                <div class="feature-item">
+                    <div>🔄</div>
+                    <h3>OpenAI API Compatible</h3>
+                    <p>Fully compatible with OpenAI API format — no client changes required</p>
+                </div>
+
+                <div class="feature-item">
+                    <div>🌊</div>
+                    <h3>Streaming Support</h3>
+                    <p>Supports real-time streaming output for a better user experience</p>
+                </div>
+
+                <div class="feature-item">
+                    <div>🔐</div>
+                    <h3>Authentication</h3>
+                    <p>API key validation to secure the service</p>
+                </div>
+
+                <div class="feature-item">
+                    <div>🛠️</div>
+                    <h3>Flexible Configuration</h3>
+                    <p>Configured via environment variables</p>
+                </div>
+
+                <div class="feature-item">
+                    <div>📝</div>
+                    <h3>Thinking Display</h3>
+                    <p>Intelligently processes and displays the model's thinking</p>
+                </div>
+
+                <div class="feature-item">
+                    <div>📊</div>
+                    <h3>Real-time Monitoring</h3>
+                    <p>Provides a web dashboard to show live forwarding and stats</p>
+                </div>
+            </div>
+        </div>
+
+        <footer>
+            <p>© 2024 ZtoApi. Powered by Deno & Z.ai GLM-4.5</p>
+        </footer>
+    </div>
+</body>
+</html>
